mupdf exception handling
needed when large TeX file is compiled and the TeX engine pauses for fetching some data, calling external programs (see `pdflatex -shell-escape`. It leaves the main pdf in a sad state. Nevertheless, G_FILE_MONITOR_EVENT_CHANGES_DONE_HINT event is already triggered.

This causes an error in turn in mupdf, which are caught in mupdf, but not from the zathua-pdf-mupdf pluging. A patch must also be applied to zathura, see other file in this directory.